feat(dilesa-ventas-expediente-tabs): Sprint 2 — Pipeline y Estado de cuenta a tab propio
Pipeline (17 fases) y Estado de cuenta (CxC) salen del cajón de Operación a
tab propio con URL. Set final = 6 tabs del expediente:
Operación · Pipeline · Cuadratura · Estado de cuenta · Documentos · Bitácora.

- Nuevas app/dilesa/ventas/[id]/{pipeline,estado-cuenta}/page.tsx consumiendo
  el VentaDetalleProvider; Operación adelgazada.
- RBAC: sub-slugs dilesa.ventas.{pipeline,estado_cuenta} en ROUTE_TO_MODULE +
  EXPECTED_DB_MODULE_SLUGS + MODULE_DEPS (deps -> operacion) + migración con
  backfill de permisos clonando de dilesa.ventas.lista.
