+++ docs/planning/dilesa-ventas-expediente-tabs.md
@@ -4,10 +4,10 @@
 **Empresas:** DILESA
 **Schemas afectados:** Principalmente UI (Next.js) — refactor de `app/dilesa/ventas/[id]/*`: nuevo `[id]/layout.tsx` + partir el monolito `[id]/page.tsx` (2,642 L) en 6 sub-rutas (`page.tsx`=Operación + `cuadratura/`, `estado-cuenta/`, `pipeline/`, `documentos/`, `bitacora/`). RBAC: `core.modulos` + `core.permisos_rol` (6 sub-slugs nuevos `dilesa.ventas.{operacion,pipeline,cuadratura,estado_cuenta,documentos,bitacora}` + backfill de permisos clonando del padre + `NOTIFY pgrst`), `lib/permissions.ts` (`ROUTE_TO_MODULE`, `HUB_PARENT_BY_ROUTE`), `lib/permissions.test.ts` (`EXPECTED_DB_MODULE_SLUGS`). Lectura de datos sin cambios.
 **Estado:** in_progress
-**Próximo hito:** Sprint 1 en [PR #948](https://github.com/beto-sudo/BSOP/pull/948) (CI verde) + migración de sub-slugs **ya aplicada a prod** (8 roles c/u = `dilesa.ventas.lista`, ledger reconciliado) — Beto: revisar Vercel preview y mergear. Luego Sprint 2: Pipeline + Estado de cuenta a tab propio + reconciliar `<CapturarFaseHeader>`.
+**Próximo hito:** Sprint 1 mergeado (#948, en prod). Sprint 2 (Pipeline + Estado de cuenta a tab propio → 6 tabs) en PR — CI + migración de los 2 sub-slugs. Tras merge, la iniciativa cubre su objetivo; queda opcional un Sprint 3 de pulido (loading por tab, deep-links a tab desde copiloto/banners, manual in-app).
 **Dueño:** Beto
 **Creada:** 2026-06-18
-**Última actualización:** 2026-06-18 (Sprint 1 completo; 6 checks verdes; migración aplicada a prod; PR #948 por revisar/mergear)
+**Última actualización:** 2026-06-18 (Sprint 1 en prod; Sprint 2 completo en local — 6 tabs, 6 checks verdes; PR S2 por revisar/mergear)
 
 > **Sucede a** [`dilesa-ventas-expediente`](dilesa-ventas-expediente.md) (cerrada — montó el workspace por venta con cabecera + tabs + copiloto). Esta iniciativa toma esos tabs (hoy `useState` sin URL) y los sube a _routed tabs_ con layout compartido, replicando el patrón ya probado en [`app/dilesa/proyectos/[id]/layout.tsx`](../../app/dilesa/proyectos/[id]/layout.tsx).
 
@@ -64,3 +64,5 @@ Set final de tabs (6): **Operación · Pipeline · Cuadratura · Estado de cuent
 - **2026-06-18** — Promovida. Diagnóstico confirmado en código (tabs `useState` en `[id]/page.tsx:346`; cabecera ya persiste en captura vía `<CapturarFaseHeader>`; molde = `proyectos/[id]/layout.tsx`). Arranca Sprint 1.
 - **2026-06-18** — **Sprint 1 completo (local, 6 checks verdes).** El monolito de 2,642 L se partió en: `components/dilesa/venta-detalle/{types,provider,ui,shell}.tsx` (cerebro = `VentaDetalleProvider` + context `useVentaDetalle`, sub-componentes y Shell) + `[id]/layout.tsx` (detecta captura: rama expediente monta provider+shell+tabs, rama captura solo persiste la barra de tabs encima del `CapturarFaseHeader` existente) + 4 páginas de tab (`page.tsx`=Operación con Pipeline+Estado de cuenta dentro, `cuadratura/`, `documentos/`, `bitacora/`), cada una con su `<RequireAccess>` fino. Decisión de arquitectura: provider en el layout (carga única, navegación entre tabs instantánea); el provider NO se monta en captura (sin doble carga). RBAC: 4 sub-slugs `dilesa.ventas.{operacion,cuadratura,documentos,bitacora}` en `ROUTE_TO_MODULE` + `EXPECTED_DB_MODULE_SLUGS` + `MODULE_DEPS` (cadena lista→operacion→{capturas,tabs}; ajustados 3 tests de `acceso-rules`/`permissions-deps` por la nueva jerarquía) + migración `20260618155211_modulos_dilesa_ventas_expediente_tabs.sql` (INSERT + backfill de permisos clonando de `dilesa.ventas.lista` — **pendiente de aplicar por Beto**). Pipeline y Estado de cuenta siguen dentro de Operación (salen en S2).
 - **2026-06-18** — **Migración aplicada a prod** (con OK de Beto) vía MCP `apply_migration`: 4 sub-slugs creados + permisos clonados de `dilesa.ventas.lista` (8 roles c/u, verificado — nadie pierde acceso). Ledger reconciliado (`20260618155211` quedó 1:1; huérfano MCP `20260618163023` marcado reverted). Drift cross-sesión preexistente (`20260618022906`/`20260618022954` = `dilesa_ventas_geometria_desglose`, de `dilesa-cuadratura-sobreprecio`) **no tocado** — requiere reconciliación de esa sesión. PR [#948](https://github.com/beto-sudo/BSOP/pull/948) listo para revisar preview + mergear.
+- **2026-06-18** — **Sprint 1 mergeado** ([#948](https://github.com/beto-sudo/BSOP/pull/948), `ea6b1cc`) — menú persistente confirmado por Beto en preview. De paso, **reconciliado el drift ajeno** (`20260618022906` ya estaba en prod — 4 columnas de geometría verificadas — marcado `applied`; huérfano `20260618022954` `reverted`): ledger de junio 18 quedó 1:1.
+- **2026-06-18** — **Sprint 2 completo (local, 6 checks verdes).** Pipeline y Estado de cuenta salen de Operación a tab propio: nuevas `pipeline/page.tsx` y `estado-cuenta/page.tsx` (con sus 3 drawers) consumiendo el context; Operación adelgazada (copiloto + PDFs + movimientos + datos + expediente digital). Set final = 6 tabs (Operación · Pipeline · Cuadratura · Estado de cuenta · Documentos · Bitácora). RBAC: 2 sub-slugs `dilesa.ventas.{pipeline,estado_cuenta}` (deps → operacion) + migración `20260618164920_modulos_dilesa_ventas_pipeline_estado_cuenta.sql`. Nota: la "reconciliación de `<CapturarFaseHeader>`" del plan original resultó **innecesaria** — en captura el layout solo pone la barra de tabs y el header pone su back-link/ficha; no hay duplicado. El único pulido (orden tabs↔back-link en captura) se difiere a un eventual Sprint 3.
